+++ src/de/omegazirkel/risingworld/adminutils/db/PrisonKey.java
@@ -1,0 +1,4 @@
+package de.omegazirkel.risingworld.adminutils.db;
+
+public record PrisonKey(long areaId) {
+}